+++ EssentialFeed/Feed Cache/Infrastructure/CoreData/ManagedCache.swift
@@ -1,0 +1,29 @@
+//
+//  ManagedCache.swift
+//  EssentialFeed
+//
+//  Created by nicaho on 2024/7/2.
+//
+
+import CoreData
+
+@objc(ManagedCache)
+internal class ManagedCache: NSManagedObject {
+    @NSManaged var timestamp: Date
+    @NSManaged var feed: NSOrderedSet
+    
+    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
+        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
+        request.returnsObjectsAsFaults = false
+        return try context.fetch(request).first
+    }
+    
+    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
+        try find(in: context).map(context.delete)
+        return ManagedCache(context: context)
+    }
+    
+    var localFeed: [LocalFeedImage] {
+        feed.compactMap { ($0 as? ManagedFeedImage)?.local }
+    }
+}
