+++ EssentialFeed/Feed Cache/Infrastructure/CoreData/ManagedFeedImage.swift
@@ -1,0 +1,32 @@
+//
+//  ManagedFeedImage.swift
+//  EssentialFeed
+//
+//  Created by nicaho on 2024/7/2.
+//
+
+import CoreData
+
+@objc(ManagedFeedImage)
+internal class ManagedFeedImage: NSManagedObject {
+    @NSManaged var id: UUID
+    @NSManaged var imageDescription: String?
+    @NSManaged var location: String?
+    @NSManaged var url: URL
+    @NSManaged var cache: ManagedCache
+    
+    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
+        NSOrderedSet(array: localFeed.map { local in
+            let managed = ManagedFeedImage(context: context)
+            managed.id = local.id
+            managed.imageDescription = local.description
+            managed.location = local.location
+            managed.url = local.url
+            return managed
+        })
+    }
+    
+    var local: LocalFeedImage {
+        LocalFeedImage(id: id, description: imageDescription, location: location, imageURL: url)
+    }
+}
